feat: Wire up game selection
