+++ frontend/src/state/game/actions.js
@@ -1,0 +1,4 @@
+export const setGame = (game) => ({
+  type: 'SET_GAME',
+  payload: { game },
+});
