+++ frontend/src/state/reducer.js
@@ -1,0 +1,5 @@
+import { combineReducers } from 'redux';
+
+import game from './game/reducer';
+
+export default combineReducers({ game });
